% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boulevard.R
\name{grow_boulevard_trees}
\alias{grow_boulevard_trees}
\title{Grow Boulevard Trees}
\usage{
grow_boulevard_trees(x, y, trees = 100L, task = "classification", ...)
}
\arguments{
\item{x}{a data X}

\item{y}{a data Y}

\item{trees}{num of trees}

\item{task}{a task}

\item{...}{other eng args}
}
\value{
a fitted model
}
\description{
Grow Boulevard Trees
}
