% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctree.R
\name{grow_conditional_trees}
\alias{grow_conditional_trees}
\title{Grow Conditional Forest}
\usage{
grow_conditional_trees(x, y, trees = 100L, task = "classification", ...)
}
\arguments{
\item{x}{a data X}

\item{y}{a data Y}

\item{trees}{num of trees}

\item{task}{a task}

\item{...}{other eng args}
}
\value{
a fitted model
}
\description{
Grow Conditional Forest
}
