% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrgboost.R
\name{grow_energy_trees}
\alias{grow_energy_trees}
\title{Grow Energy-Based Generative Trees}
\usage{
grow_energy_trees(x, y, trees = 100L, task = "classification", ...)
}
\arguments{
\item{x}{a data X}

\item{y}{a data Y}

\item{trees}{num of trees}

\item{task}{a task}

\item{...}{other eng args}
}
\value{
a fitted model
}
\description{
Grow Energy-Based Generative Trees
}
