% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evtree.R
\name{grow_evolutionary_trees}
\alias{grow_evolutionary_trees}
\title{Grow Evolutionary Trees}
\usage{
grow_evolutionary_trees(x, y, trees = 100L, task = "classification", ...)
}
\arguments{
\item{x}{a data X}

\item{y}{a data Y}

\item{trees}{num of trees}

\item{task}{a task}

\item{...}{other eng args}
}
\value{
a fitted model
}
\description{
Grow Evolutionary Trees
}
