% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebm.R
\name{grow_explainable_trees}
\alias{grow_explainable_trees}
\title{Grow Explainable Trees}
\usage{
grow_explainable_trees(
  x,
  y,
  trees = 50000L,
  early_stop_tree = 100L,
  task = "classification",
  ...
)
}
\arguments{
\item{x}{a data X}

\item{y}{a data Y}

\item{trees}{max num of trees (default: 50000)}

\item{early_stop_tree}{trigger early stopping if no improvement (default: 100)}

\item{task}{a task}

\item{...}{other eng args}
}
\value{
a fitted model
}
\description{
Grow Explainable Trees
}
