% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairgbm.R
\name{grow_fair_trees}
\alias{grow_fair_trees}
\title{Grow Fair Trees}
\usage{
grow_fair_trees(x, y, s, trees = 100L, task = "classification", ...)
}
\arguments{
\item{x}{a data X}

\item{y}{a data Y}

\item{s}{a contraint group sensitivity for fairness}

\item{trees}{num of trees}

\item{task}{a task}

\item{...}{other eng args}
}
\value{
a fitted fairgbm model
}
\description{
Grow Fair Trees
}
