% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grande.R
\name{grow_grande_trees}
\alias{grow_grande_trees}
\title{Grow GRANDE Trees}
\usage{
grow_grande_trees(x, y, trees = 100L, task = "classification", ...)
}
\arguments{
\item{x}{a data X}

\item{y}{a data Y}

\item{trees}{num of trees}

\item{task}{a task}

\item{...}{other eng args}
}
\value{
a fitted model
}
\description{
Grow GRANDE Trees
}
