% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktboost.R
\name{grow_kernel_trees}
\alias{grow_kernel_trees}
\title{Grow kernel-Based Generative Trees}
\usage{
grow_kernel_trees(x, y, trees = 100L, task = "classification", ...)
}
\arguments{
\item{x}{a data X}

\item{y}{a data Y}

\item{trees}{num of trees}

\item{task}{a task}

\item{...}{other eng args}
}
\value{
a fitted model
}
\description{
Grow kernel-Based Generative Trees
}
