% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphboost.R
\name{grow_morph_trees}
\alias{grow_morph_trees}
\title{Grow Morph Trees}
\usage{
grow_morph_trees(x, y, trees = 100L, task = "classification", ...)
}
\arguments{
\item{x}{a data X}

\item{y}{a data Y}

\item{trees}{num of trees}

\item{task}{a task}

\item{...}{other eng args}
}
\value{
a fitted model
}
\description{
Grow Morph Trees
}
