% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngboost.R
\name{grow_natural_trees}
\alias{grow_natural_trees}
\title{Grow Natural Gradient Trees}
\usage{
grow_natural_trees(x, y, trees = 100L, task = "classification", ...)
}
\arguments{
\item{x}{a data X}

\item{y}{a data Y}

\item{trees}{num of trees}

\item{task}{a task}

\item{...}{other eng args}
}
\value{
a fitted model
}
\description{
Grow Natural Gradient Trees
}
