% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perpetual.R
\name{grow_perpetual_trees}
\alias{grow_perpetual_trees}
\title{Grow perpetual Trees}
\usage{
grow_perpetual_trees(x, y, budget = 0.5, task = "classification", ...)
}
\arguments{
\item{x}{a data X}

\item{y}{a data Y}

\item{budget}{the 'predictive power' of perpetual booster}

\item{task}{a task}

\item{...}{other eng args}
}
\value{
a fitted model
}
\description{
Grow perpetual Trees
}
