% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapboost.R
\name{grow_snap_trees}
\alias{grow_snap_trees}
\title{Grow Snap Trees}
\usage{
grow_snap_trees(x, y, trees = 100L, task = "classification", ...)
}
\arguments{
\item{x}{a data X}

\item{y}{a data Y}

\item{trees}{num of trees}

\item{task}{a task}

\item{...}{other eng args}
}
\value{
a fitted model
}
\description{
Grow Snap Trees
}
