% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildwood.R
\name{grow_wild_trees}
\alias{grow_wild_trees}
\title{Grow Wild Trees}
\usage{
grow_wild_trees(x, y, trees = 100L, task = "classification", ...)
}
\arguments{
\item{x}{a data X}

\item{y}{a data Y}

\item{trees}{num of trees}

\item{task}{a task}

\item{...}{other eng args}
}
\value{
a fitted model
}
\description{
Grow Wild Trees
}
