% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ydf.R
\name{grow_yggdrasil_trees}
\alias{grow_yggdrasil_trees}
\title{Grow Yggdrasil Trees}
\usage{
grow_yggdrasil_trees(x, y, task = "classification", trees = 300L, ...)
}
\arguments{
\item{x}{a data X}

\item{y}{a data Y}

\item{task}{a task}

\item{trees}{num of trees}

\item{...}{other eng args}
}
\value{
a fitted model
}
\description{
Grow Yggdrasil Trees
}
