% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boulevard.R
\name{harvest_boulevard_trees}
\alias{harvest_boulevard_trees}
\title{Harvest Boulevard Trees}
\usage{
harvest_boulevard_trees(fit, x, ...)
}
\arguments{
\item{fit}{a fitted boulevard model}

\item{x}{a set of predictors}

\item{...}{placeholder}
}
\value{
predictions
}
\description{
Harvest Boulevard Trees
}
