% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctree.R
\name{harvest_conditional_trees}
\alias{harvest_conditional_trees}
\title{Harvest Conditional Forest}
\usage{
harvest_conditional_trees(fit, x, ...)
}
\arguments{
\item{fit}{a fitted evtree model}

\item{x}{a set of predictors}

\item{...}{placeholder}
}
\value{
predictions
}
\description{
Harvest Conditional Forest
}
