% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrgboost.R
\name{harvest_energy_trees}
\alias{harvest_energy_trees}
\title{Harvest Energy-Based Generative Trees}
\usage{
harvest_energy_trees(fit, x, ...)
}
\arguments{
\item{fit}{a fitted nrgboost model}

\item{x}{a set of predictors}

\item{...}{placeholder}
}
\value{
predictions
}
\description{
Harvest Energy-Based Generative Trees
}
