% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evtree.R
\name{harvest_evolutionary_trees}
\alias{harvest_evolutionary_trees}
\title{Harvest Evolutionary Trees}
\usage{
harvest_evolutionary_trees(fit, x, ...)
}
\arguments{
\item{fit}{a fitted evtree model}

\item{x}{a set of predictors}

\item{...}{placeholder}
}
\value{
predictions
}
\description{
Harvest Evolutionary Trees
}
