% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebm.R
\name{harvest_explainable_trees}
\alias{harvest_explainable_trees}
\title{Harvest Wild Trees}
\usage{
harvest_explainable_trees(fit, x, ...)
}
\arguments{
\item{fit}{a fitted wildwood model}

\item{x}{a set of predictors}

\item{...}{placeholder}
}
\value{
predictions
}
\description{
Harvest Wild Trees
}
