% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme.R
\name{harvest_extreme_trees}
\alias{harvest_extreme_trees}
\title{Harvest extreme Trees}
\usage{
harvest_extreme_trees(fit, x, ...)
}
\arguments{
\item{fit}{a fitted extreme model}

\item{x}{a set of predictors}

\item{...}{placeholder}
}
\value{
predictions
}
\description{
Harvest extreme Trees
}
