% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairgbm.R
\name{harvest_fair_trees}
\alias{harvest_fair_trees}
\title{Harvest Fair Trees}
\usage{
harvest_fair_trees(fit, x, ...)
}
\arguments{
\item{fit}{a fitted fairgbm model}

\item{x}{a set of predictors}

\item{...}{placeholder}
}
\value{
predictions
}
\description{
Harvest Fair Trees
}
