% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grande.R
\name{harvest_grande_trees}
\alias{harvest_grande_trees}
\title{Harvest GRANDE Trees}
\usage{
harvest_grande_trees(fit, x, ...)
}
\arguments{
\item{fit}{a fitted ngboost model}

\item{x}{a set of predictors}

\item{...}{placeholder}
}
\value{
predictions
}
\description{
Harvest GRANDE Trees
}
