% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktboost.R
\name{harvest_kernel_trees}
\alias{harvest_kernel_trees}
\title{Harvest kernel-Based Generative Trees}
\usage{
harvest_kernel_trees(fit, x, ...)
}
\arguments{
\item{fit}{a fitted KTboost model}

\item{x}{a set of predictors}

\item{...}{placeholder}
}
\value{
predictions
}
\description{
Harvest kernel-Based Generative Trees
}
