% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langevin.R
\name{harvest_langevin_trees}
\alias{harvest_langevin_trees}
\title{Harvest Langevin Trees}
\usage{
harvest_langevin_trees(fit, x, ...)
}
\arguments{
\item{fit}{a fitted langevin model}

\item{x}{a set of predictors}

\item{...}{placeholder}
}
\value{
predictions
}
\description{
Harvest Langevin Trees
}
