% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{harvest_linear_trees}
\alias{harvest_linear_trees}
\title{Harvest Linear Trees}
\usage{
harvest_linear_trees(fit, x, ...)
}
\arguments{
\item{fit}{a fitted linear dart model}

\item{x}{a set of predictors}

\item{...}{placeholder}
}
\value{
predictions
}
\description{
Harvest Linear Trees
}
