% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphboost.R
\name{harvest_morph_trees}
\alias{harvest_morph_trees}
\title{Harvest Morph Trees}
\usage{
harvest_morph_trees(fit, x, ...)
}
\arguments{
\item{fit}{a fitted morphboost model}

\item{x}{a set of predictors}

\item{...}{placeholder}
}
\value{
predictions
}
\description{
Harvest Morph Trees
}
