% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngboost.R
\name{harvest_natural_trees}
\alias{harvest_natural_trees}
\title{Harvest Natural Gradient Trees}
\usage{
harvest_natural_trees(fit, x, ...)
}
\arguments{
\item{fit}{a fitted ngboost model}

\item{x}{a set of predictors}

\item{...}{placeholder}
}
\value{
predictions
}
\description{
Harvest Natural Gradient Trees
}
