% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perpetual.R
\name{harvest_perpetual_trees}
\alias{harvest_perpetual_trees}
\title{Harvest Perpetual Trees}
\usage{
harvest_perpetual_trees(fit, x, ...)
}
\arguments{
\item{fit}{a fitted perpetual model}

\item{x}{a set of predictors}

\item{...}{placeholder}
}
\value{
predictions
}
\description{
Harvest Perpetual Trees
}
