% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapboost.R
\name{harvest_snap_trees}
\alias{harvest_snap_trees}
\title{Harvest Snap Trees}
\usage{
harvest_snap_trees(fit, x, ...)
}
\arguments{
\item{fit}{a fitted snapboost model}

\item{x}{a set of predictors}

\item{...}{placeholder}
}
\value{
predictions
}
\description{
Harvest Snap Trees
}
