% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildwood.R
\name{harvest_wild_trees}
\alias{harvest_wild_trees}
\title{Harvest Wild Trees}
\usage{
harvest_wild_trees(fit, x, ...)
}
\arguments{
\item{fit}{a fitted wildwood model}

\item{x}{a set of predictors}

\item{...}{placeholder}
}
\value{
predictions
}
\description{
Harvest Wild Trees
}
