% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ydf.R
\name{harvest_yggdrasil_trees}
\alias{harvest_yggdrasil_trees}
\title{Harvest Yggdrasil Trees}
\usage{
harvest_yggdrasil_trees(fit, x, ...)
}
\arguments{
\item{fit}{a fitted snapboost model}

\item{x}{a set of predictors}

\item{...}{placeholder}
}
\value{
predictions
}
\description{
Harvest Yggdrasil Trees
}
