% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daisugi-package.R
\docType{package}
\name{daisugi-package}
\alias{daisugi}
\alias{daisugi-package}
\title{daisugi: Novel Tree Machines for Classification and Regression}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

daisugi provides an interface for not-so-well-known tree machines. particularly focused on novel tree and boosting algorithms for machine learning tasks. those tasks being tabular regression and classification problems. daisugi provides a unified interface to YDF, SnapBoost, WildWood, Perpetual, and more.
}
\author{
\strong{Maintainer}: frank hull \email{frankiethull@proton.me}

}
\keyword{internal}
