% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daisugi-package.R
\docType{package}
\name{daisugi-package}
\alias{daisugi}
\alias{daisugi-package}
\title{daisugi: Novel Tree Machines for Classification and Regression}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

daisugi provides a unified interface for specialized and novel tree-based machine learning algorithms. The package focuses on tabular classification and regression tasks, offering streamlined access to "not-so-well-known" architectures. Users can implement and compare various backends through a consistent API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://frankiethull.github.io/daisugi/}
  \item \url{https://github.com/frankiethull/daisugi}
  \item Report bugs at \url{https://github.com/frankiethull/daisugi/issues}
}

}
\author{
\strong{Maintainer}: frank hull \email{frankiethull@proton.me} [copyright holder]

}
\keyword{internal}
